+++ lib/ffi-gobject_introspection/i_callable_info.rb
@@ -31,5 +31,13 @@ module GObjectIntrospection
     def skip_return?
       Lib.g_callable_info_skip_return @gobj
     end
+
+    def instance_ownership_transfer
+      if VERSION >= "1.42"
+        Lib.g_callable_info_get_instance_ownership_transfer @gobj
+      else
+        :nothing
+      end
+    end
   end
 end