+++ lib/gir_ffi/builders/function_builder.rb
@@ -73,8 +73,16 @@ module GirFFI
 
       def function_call_arguments
         ca = @argument_builder_collection.call_argument_names
-        ca.unshift 'self' if @info.method?
+        ca.unshift receiver_call_argument if @info.method?
         ca
+      end
+
+      def receiver_call_argument
+        if @info.instance_ownership_transfer == :everything
+          'self.ref'
+        else
+          'self'
+        end
       end
     end
   end